# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protomind API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/protocols/{id}/documents/pdf": {
            "get": {
                "tags": [
                    "ProtocolDocuments"
                ],
                "summary": "Сгенерировать PDF протокола",
                "description": "Метод для генерации PDF версии протокола",
                "operationId": "generatePdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Протокол не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/protocols/{id}/documents/docx": {
            "get": {
                "tags": [
                    "ProtocolDocuments"
                ],
                "summary": "Сгенерировать DOCX протокола",
                "description": "Метод для генерации DOCX версии протокола",
                "operationId": "generateDocx",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Протокол не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/protocols/{protocolId}/tasks": {
            "post": {
                "tags": [
                    "ProtocolTasks"
                ],
                "summary": "Создать новую задачу в протоколе",
                "description": "Метод для создания новой задачи в протоколе",
                "operationId": "storeProtocolTask",
                "parameters": [
                    {
                        "name": "protocolId",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProtocolTaskStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProtocolTaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Протокол не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Протокол не найден."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/protocols/{protocolId}/tasks/{taskId}": {
            "put": {
                "tags": [
                    "ProtocolTasks"
                ],
                "summary": "Обновить задачу в протоколе",
                "description": "Метод для обновления задачи в протоколе",
                "operationId": "updateProtocolTask",
                "parameters": [
                    {
                        "name": "protocolId",
                        "in": "path",
                        "description": "ID протокола",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProtocolTaskUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProtocolTaskResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Задача не найдена."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProtocolTaskStoreRequest": {
                "title": "Запрос на создание задачи протокола",
                "description": "Запрос, содержащий данные для создания задачи протокола",
                "properties": {
                    "responsible_id": {
                        "description": "ID ответственного пользователя",
                        "type": "integer"
                    },
                    "essence": {
                        "description": "Суть задачи",
                        "type": "string",
                        "maxLength": 65000
                    },
                    "deadline": {
                        "description": "Срок выполнения задачи",
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-01"
                    }
                },
                "type": "object"
            },
            "ProtocolTaskUpdateRequest": {
                "title": "Запрос на обновление задачи протокола",
                "description": "Запрос, содержащий данные для обновления задачи протокола",
                "properties": {
                    "responsible_id": {
                        "description": "ID ответственного пользователя",
                        "type": "integer",
                        "nullable": true
                    },
                    "essence": {
                        "description": "Суть задачи",
                        "type": "string",
                        "maxLength": 65000,
                        "nullable": true
                    },
                    "deadline": {
                        "description": "Срок выполнения задачи",
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-01",
                        "nullable": true
                    },
                    "status": {
                        "description": "Статус задачи",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ProtocolResource": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Protocol"
                    }
                ]
            },
            "ProtocolTaskResource": {
                "title": "Задача протокола",
                "description": "Ресурс задачи протокола",
                "properties": {
                    "id": {
                        "description": "ID задачи протокола",
                        "type": "integer"
                    },
                    "protocol": {
                        "$ref": "#/components/schemas/ProtocolResource"
                    },
                    "responsible": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "essence": {
                        "description": "Суть задачи",
                        "type": "string"
                    },
                    "deadline": {
                        "description": "Срок выполнения задачи",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Статус задачи",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "Пользователь",
                "description": "Ресурс пользователя",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "full_name": {
                        "description": "Полное имя пользователя",
                        "type": "string"
                    },
                    "email": {
                        "description": "Электронная почта пользователя",
                        "type": "string",
                        "format": "email"
                    },
                    "login": {
                        "description": "Логин пользователя",
                        "type": "string"
                    },
                    "role": {
                        "description": "Роль пользователя",
                        "type": "string"
                    },
                    "department": {
                        "description": "Отдел пользователя",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата и время создания пользователя",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления пользователя",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Protocol": {
                "title": "Протокол",
                "description": "Модель протокола",
                "properties": {
                    "id": {
                        "description": "ID протокола",
                        "type": "integer"
                    },
                    "theme": {
                        "description": "Тема протокола",
                        "type": "string"
                    },
                    "agenda": {
                        "description": "Повестка дня",
                        "type": "string"
                    },
                    "secretary_id": {
                        "description": "ID секретаря",
                        "type": "integer"
                    },
                    "director_id": {
                        "description": "ID директора",
                        "type": "integer"
                    },
                    "event_date": {
                        "description": "Дата события",
                        "type": "string",
                        "format": "date"
                    },
                    "status": {
                        "description": "Статус протокола",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата и время создания протокола",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления протокола",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}